Minor refactoring and extraction of Interfaces from service implementations

+++ src/main/java/no/ntnu/appdevapi/DAO/RatingRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 /**
  * Repository interface for ratings.
  */
-public interface RatingRepository extends CrudRepository<Rating, Integer> {
+public interface RatingRepository extends CrudRepository<Rating, Long> {
     List<Rating> findByProduct(Product product);
 }
